requirements.txt
Django>=4.0.4
djangorestframework>=3.13.1, <3.14.0
psycopg2>=2.9.3
drf-spectacular>=0.22.1
coverage>=6.3.3
